Tunnel Ansatz 0.4.0 [+ Rechte hinzugefügt ; + Dokumentation und Kommentare hinzugefügt] (#70)
* + Tunneling-Ansatz ;
+ server.js tunneling ver.
+ index.html tunneling ver.
+ live-cache.js tunniling ver.

* + Member Hinzufügen;
+ Generiert einen zufälligen Identifier für jeden User ;
+ Zeigt eine Liste aller Benutzer an;
+ Neuer Cache in den Members gepseichert werden |
* Daten werden nun beim erstellen einer neuen Session gereinigt ;

* + Löschen von Gruppen ;
+ Sessions können nun gelöscht werden |
* Der Identifikator von Clients ändert sich nun nicht mehr mit jeder Session

* + Dokumentation und Komentare hinzugefügt ;
* Kommentare für Funktionen verbessert / hinzugefügt

* + AddPermission funktion |
* Der ersteller der Session erhält automatisch rechte

* + Rechte hinzugefügt ;
+ Der Host einer Gruppe erhält automatisch Rechte ;
+ Funktion hinzugefügt Nutzern Rechte zu geben ;
+ Der Code überprüft bei kritischen Anfragen, ob der Nutzer Rechte hat ;
+ convertToPrint in memberCache |
* Nutzernamen wird nun eine Zahl angehängt, wenn dieser bereits vergeben ist ;
* HTML mit etwas CSS aufgräumt ;
* Code verwendet nun einheitlich die bezeichnung sessionKey statt cacheKay

+++ Rapid-Prototype/tunneling-Ansatz/server.js
@@ -1,4 +1,4 @@
-// Ursprungscode und Konzept von: https://technology.amis.nl/cloud/implementing-serverless-multi-client-session-synchronization-with-oracle-cloud-infrastructure/
+/* Konzept von: https://technology.amis.nl/cloud/implementing-serverless-multi-client-session-synchronization-with-oracle-cloud-infrastructure/ */
 const cache = require('./live-cache.js');
 const memberCache = require('./member-cache.js');
 const express = require('express');
@@ -9,8 +9,7 @@ const cors = require('cors');
 
 const app = express();
 const port = 3000;
-// Header für CORS Responses
-const allowedHeader = {
+const allowedHeader = {     // Header für CORS Responses
     'Access-Control-Allow-Origin': '*',
     'Access-Control-Allow-Methods': 'GET,OPTIONS,POST,PUT',
     'Access-Control-Allow-Headers': '*',
@@ -21,6 +20,7 @@ console.log("Root Path: " + rootPath)
 
 app.use(cors());
 app.use(bodyParser.json());
+app.use('/src/css/main.css', express.static(path.join(__dirname, 'src/css/main.css')));
 
 // Set up localtunnel
 const subdomain = 'ripe-mangos-super-smell';
@@ -28,18 +28,17 @@ localtunnel(port, { subdomain: subdomain }, (err, tunnel) => {
     if (err) {
         console.error('Error creating tunnel:', err);
     } else {
-        console.log(`Tunnel URL: ${tunnel.url}`);
-        // Share the tunnel URL with others
+        console.log(`Tunnel URL: ${tunnel.url}`); // Public Url
     }
 });
 
+// Gibt den SessionContext aus wenn ein sessionKey vorhanden ist, ansonsten gibt die HTML Seite
 app.get('/', (req, res) => {
     const sessionKey = req.query.sessionKey;
 
     if (!sessionKey || sessionKey.trim() === '') {
         res.redirect('index.html');
     } else {
-        //console.log("Deine Mudda war hier")
         result = cache.readFromCache(sessionKey);
         member = memberCache.readFromCache(sessionKey);
         console.log("Result: " + JSON.stringify(result))
@@ -53,9 +52,14 @@ app.get('/', (req, res) => {
     }
 });
 
+// Lässt einen neuen sessionKey für eine neue Session generieren und legt eine Gruppe für diese an
 app.post('/', (req, res) => {
     result = cache.startNewSession();
+    //memberCache.initializeCache(result.sessionKey);
     console.log("Result: " + JSON.stringify(result))
+    const print = req.body;
+    console.log("Das ist der Host: " + JSON.stringify(print.fingerprint))
+    memberCache.addHost("\"" + result.sessionKey + "\"", print.fingerprint)
     res.set(allowedHeader);
 
     res.status(200).send({
@@ -63,6 +67,7 @@ app.post('/', (req, res) => {
     });
 });
 
+// Updated die Daten der Session mit dem übermittelten sessionKey
 app.put('/', (req, res) => {
     const theBody = req.body;
     JSON.stringify("The Body: " + theBody);
@@ -78,18 +83,24 @@ app.put('/', (req, res) => {
     });
 });
 
+// löscht die Daten der Session mit dem übermittelten sessionKey
 app.delete('/', (req, res) => {
     const sessionKey = req.query.sessionKey;
     const print = req.query.print;
 
-    result = cache.deleteSession(sessionKey)
-    memberResult = memberCache.deleteSession(sessionKey)
-    console.log("Result von Cache delte: " + JSON.stringify(result) + " und Member Cache: " + JSON.stringify(memberResult))
-    res.set(allowedHeader);
+    if (memberCache.hasPermission(sessionKey, print)) {
+        result = cache.deleteSession(sessionKey)
+        memberResult = memberCache.deleteSession(sessionKey)
+        console.log("Result von Cache delte: " + JSON.stringify(result) + " und Member Cache: " + JSON.stringify(memberResult))
+        res.set(allowedHeader);
 
-    res.status(200).send();
+        res.status(200).send();
+    } else {
+        res.status(403).send("Sie haben keine Rechte für diese Gruppe");
+    }
 });
 
+// Funktion für Test und Debbunging zwecke. (CORS macht manchmal Options Requests um zu prüfen ob die Anfrage gültig ist)
 app.options('/', (req, res) => {
     console.log("Das Kitzelt");
 
@@ -97,18 +108,43 @@ app.options('/', (req, res) => {
     res.status(200).send();
 })
 
+// fügt ein Member einer Session Gruppe hinzu
 app.post('/member', async (req, res) => {
     try {
         const sessionKey = req.query.sessionKey;
         const member = req.query.member;
         const print = req.body;
-        
+
         console.log("Die VisitorId: " + JSON.stringify(print.fingerprint))
         result = memberCache.writeToCache(sessionKey, print.fingerprint, member);
         console.log("Result: " + JSON.stringify(result));
 
         res.set(allowedHeader);
         res.status(200).send();
+    } catch (error) {
+        console.error("Error:", error);
+        res.status(500).send("Internal Server Error");
+    }
+});
+
+// gibt einem Member Rechte
+app.post('/permission', async (req, res) => {
+    try {
+        const sessionKey = req.query.sessionKey;
+        const member = req.query.member;
+        const print = req.body;
+
+        console.log("Der permission Member: " + JSON.stringify(member))
+        console.log("Der permission print: " + JSON.stringify(print.fingerprint))
+        if (memberCache.hasPermission(sessionKey, print.fingerprint)) {
+            result = memberCache.addPermission(sessionKey, member);
+            console.log("Permission Result: " + JSON.stringify(result));
+
+            res.set(allowedHeader);
+            res.status(200).send();
+        } else {
+            res.status(403).send("Sie haben keine Rechte für diese Gruppe");
+        }
     } catch (error) {
         console.error("Error:", error);
         res.status(500).send("Internal Server Error");
